- Memperbaiki bug - bug UI kecil - Membuat aplikasi full menggunakan backend untuk autentikasi - Mengganti fungsi DialogRecipeTutorialAdapter untuk menampilkan MealHistory harian di HomeMealHistoryFragment - Menambahkan HomeMealHistoryFragment dan RecipeDetailActivity - Menambahkan TransitionProgramActivity untuk menampilkan hasil dari program ketika program selesai dijalani

+++ app/src/main/java/com/telyu/nourimate/data/local/models/Detail.kt
@@ -1,7 +1,6 @@
 package com.telyu.nourimate.data.local.models
 
 import android.os.Parcelable
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.parcelize.Parcelize
@@ -13,19 +12,12 @@ data class Detail(
     @PrimaryKey(autoGenerate = true)
     val detailId: Int,
     val dob: Date?,
-
     val height: Float?,
-
     val weight: Float?,
-
     val waistSize: Float?,
-
     val gender: String,
-
     val allergen: String,
-
     val disease: String,
-
-    val bmi: Int?,
+    val bmi: Float?,
 
     ): Parcelable
